### 2024-03-13 #6475 @editorialbot
[PRE REVIEW]: JAXbind: Bind any function to JAX
**Submitting author:** @roth-jakob (<a href="http://orcid.org/0000-0002-8873-8215">Jakob Roth</a>)
**Repository:** https://github.com/NIFTy-PPL/JAXbind
**Branch with paper.md** (empty if default branch):
**Version:** 0.1.0
