### 2024-03-13 #6475 @danielskatz
👋 @roth-jakob - thanks for your submission.

Could you work on the possibly missing DOI that editorialbot suggests, but note that this may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
