### 2024-03-13 c74ba90
paper.bib: add dois
