+++ paper/paper.bib
@@ -57,7 +57,8 @@
   numpages  = {18},
   keywords  = {automatic differentiation, tasks, OpenMP, compiler, Julia, parallel, Enzyme, C++, RAJA, hybrid parallelization, MPI, distributed, LLVM},
   location  = {Dallas, Texas},
-  series    = {SC '22}
+  series    = {SC '22},
+  doi       = {10.1109/SC41404.2022.00065}
 }
 @software{Resolve2024,
   author = {Arras, Philipp and Roth, Jakob and Ding, Simon and Reinecke, Martin and Fuchs, Richard and Johnson, Vishal},
@@ -72,5 +73,6 @@
   year          = 2024,
   eprint        = {2402.16683},
   archiveprefix = {arXiv},
-  primaryclass  = {astro-ph.IM}
+  primaryclass  = {astro-ph.IM},
+  doi           = {10.48550/arXiv.2402.16683},
 }
