### 2024-03-13 #6475 @roth-jakob
Thanks for instantly opening the review issue!

I added DOIs wherever possible. Some of the citations refer to software or proceedings and don't have a DOI.
