### 2024-03-14 #6475 @roth-jakob
Thank you for the update!

We are aware that the codebase of JAXbind is relatively small. Nevertheless, we believe it to be highly useful for various applications requiring custom functions in JAX. Binding custom functions to JAX without our JAXbind package is a big effort, as described for one example function in Dan Foreman-Mackey's repository: https://github.com/dfm/extending-jax

We are very happy that we managed to code general JAX bindings in a compact form, but of course, understand you have to check the JOSS guidelines. Regarding the age of the repository and the number of commits, please be aware that we previously worked on this project in other repositories. See, for example, the various branches in: https://gitlab.mpcdf.mpg.de/jroth/extending-jax-and-nifty

Great to hear that Dan Foreman-Mackey agreed to review!
