### 2024-03-22 #6475 @danielskatz
👋 @wsmoses & @hawkinsp - would one or both you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
