+++ jaxbind/__init__.py
@@ -1,4 +1,5 @@
 __version__ = "1.0.0"
 
+from . import contrib
 from .jaxbind import get_linear_call, get_nonlinear_call
 from .misc import load_kwargs
