+++ jaxbind/contrib/jaxducc0.py
@@ -18,3 +18,17 @@ def _fht_abstract(*args, **kwargs):
 
 
 genuine_fht = get_linear_call(_fht, _fht, _fht_abstract, _fht_abstract)
+
+
+def _c2c(out, args, kwargs_dump):
+    (x,) = args
+    kwargs = load_kwargs(kwargs_dump)
+    ducc0.fft.c2c(x, out=out[0], **kwargs)
+
+
+def _c2c_abstract(*args, **kwargs):
+    (x,) = args
+    return ((x.shape, x.dtype),)
+
+
+c2c = get_linear_call(_c2c, _c2c, _c2c_abstract, _c2c_abstract)
