### 2024-03-22 14e35d2
jaxducc0: Implement SHT for HEALPix stub using Martin's test
