### 2024-03-22 #6475 @danielskatz
Great - thanks @wsmoses!  I'll add you and we'll start.  JOSS reviews are typically expected in about 4 weeks, but since they are iterative, this really depends on what reviewers think needs to be done and how the authors respond. The authors can work on items as soon as reviewers raise them, rather than waiting for a review to be fully completed.
