### 2024-03-22 #6532 @editorialbot
[REVIEW]: JAXbind: Bind any function to JAX
**Submitting author:** @roth-jakob (<a href="http://orcid.org/0000-0002-8873-8215">Jakob Roth</a>)
**Repository:** https://github.com/NIFTy-PPL/JAXbind
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
