### 2024-03-23 b3ee43c
test_contrib_jaxducc0: test wgridder bindings
