### 2024-03-23 865b760
jaxducc0: write wgridder result directly into out
