### 2024-03-23 279528b
jaxducc0: allow for additional kwargs to wgridder
