### 2024-03-23 3292a39
test_contrib_jaxducc0: test_wgridder use functools partial
