### 2024-03-25 288176e
test_batching.py: Format
