### 2024-03-25 2b11c40
jaxducc0: Copy over doc-strings for c2c and genuine_fht
