### 2024-03-25 fefeab9
jaxducc0: Implement and test native batching
