+++ test/test_contrib_jaxducc0.py
@@ -193,7 +193,9 @@ def test_healpix(lmmax, nside, spin, dtype, nthreads):
     hpp = jaxducc0.get_healpix_sht(nside, lmax, mmax, spin, nthreads)
 
     map0 = np.array(hpp(alm0r)[0])
-    map1 = ducc0.sht.synthesis(alm=alm0, lmax=lmax, mmax=mmax, spin=spin, nthreads=nthreads, **ducc0.healpix.Healpix_Base(nside, "RING").sht_info())
+    map1 = ducc0.sht.synthesis(
+        alm=alm0, lmax=lmax, mmax=mmax, spin=spin, nthreads=nthreads,
+        **ducc0.healpix.Healpix_Base(nside, "RING").sht_info())
     np.testing.assert_allclose(map0, map1, atol=1e-5, rtol=1e-5)
 
     max_order = 2
