+++ test/test_contrib_jaxducc0.py
@@ -194,12 +194,18 @@ def test_healpix(lmmax, nside, spin, dtype, nthreads):
 
     map0 = np.array(hpp(alm0r)[0])
     map1 = ducc0.sht.synthesis(
-        alm=alm0, lmax=lmax, mmax=mmax, spin=spin, nthreads=nthreads,
-        **ducc0.healpix.Healpix_Base(nside, "RING").sht_info())
+        alm=alm0,
+        lmax=lmax,
+        mmax=mmax,
+        spin=spin,
+        nthreads=nthreads,
+        **ducc0.healpix.Healpix_Base(nside, "RING").sht_info(),
+    )
     np.testing.assert_allclose(map0, map1, atol=1e-5, rtol=1e-5)
 
     max_order = 2
     check_grads(hpp, (alm0r,), order=max_order, modes=("fwd", "rev"), eps=1.0)
+
 
 #    map0 = (rng.random((ncomp, 12 * nside**2)) - 0.5).astype(dtype)
 #    hpp_t = jax.linear_transpose(hpp)
