### 2024-03-29 69df114
Fix type of cotangent in README (and paper.md)
Resolves #9 .
