### 2024-03-29 NIFTy-PPL/JAXbind#10 @dfm
[JOSS 6532] Relationship to JAX's built in "external callback" functionality
If I understand the implementation properly (although the docs are not very clear about the implementation details), this library works by [acquiring the Python GIL](https://github.com/NIFTy-PPL/JAXbind/blob/8ca830678bbed3c3fd3809af3c7de08bae0aa53e/src/_jaxbind.cc#L46C7-L46C25) and executing a [Python callback](https://github.com/NIFTy-PPL/JAXbind/blob/8ca830678bbed3c3fd3809af3c7de08bae0aa53e/src/_jaxbind.cc#L103) from within the XLA custom call. As far as I can tell, this is the same behavior offered by [JAX's built in external callback functionality](https://jax.readthedocs.io/en/latest/notebooks/external_callbacks.html), which can be combined with the [`jax.custom_jvp`](https://jax.readthedocs.io/en/latest/_autosummary/jax.custom_jvp.html) and/or [`jax.custom_vjp`](https://jax.readthedocs.io/en/latest/_autosummary/jax.custom_vjp.html) decorators to achieve the same results as JAXBind. I'd be interested to understand the difference between these two approaches, and a clear comparison should be added to the docs (and JOSS paper).

As a smaller side note, it might be also interesting to compare to [the Pallas submodule](https://jax.readthedocs.io/en/latest/pallas/index.html), which takes a slightly different approach to similar ends on GPUs and TPUs.

ref: https://github.com/openjournals/joss-reviews/issues/6532
