### 2024-03-29 NIFTy-PPL/JAXbind#12 @dfm
[JOSS 6532] Expose API docs
While the public interface functions do seem to have good docstrings, and there are nice examples of usage, the API docs are not sufficiently accessible by JOSS standards.

Related to #8

ref: https://github.com/openjournals/joss-reviews/issues/6532
