### 2024-03-29 NIFTy-PPL/JAXbind#9 @Edenhofer
Thanks for the report! The co-tangent `cotan` needs to be a list instead of a tuple because the output always is a list. #11 updates the README accordingly.
