### 2024-03-29 NIFTy-PPL/JAXbind#9 @roth-jakob
Thanks for noticing that!
