### 2024-03-29 NIFTy-PPL/JAXbind#9 @dfm
I can confirm that #11 fixes my issue.
