### 2024-03-29 NIFTy-PPL/JAXbind#10 @Edenhofer
Good point! We should definitely compare to both! I'll add a discuss of the differences to the documentation and the paper!

Yes, we briefly need to acquire the GIL to call back into python as to allow for a convenient python user-interface. For any reasonably expensive call, the time for which the C++ code will hold the GIL will in all likelihood be completely negligible. In the python call, the GIL will likely again be released very quickly assuming most codes in python will pretty quickly go back to C/C++/Rust/... (e.g. by calling scipy/numpy).

The going-back-to-python part is indeed very similar to the host-callback in JAX which can be combined with `jax.custom_jvp` **OR** `jax.custom_vjp` (["Forward-mode autodiff cannot be used on the jax.custom_vjp function and will raise an error"](https://jax.readthedocs.io/en/latest/notebooks/Custom_derivative_rules_for_Python_code.html)). The "or" is exactly what puts it apart from these functions. JAXbind not only registers a custom JVP but also a custom VJP meaning that we can e.g. access the Fisher metric of a model with a call bound using JAXbind without relying on JAX's transposition backend. For linear functions this means that we can differentiate them arbitrary many times and transpose them at will.
