### 2024-03-29 2bec9ad
jaxbind: document JAXbind internal api
