### 2024-03-29 NIFTy-PPL/JAXbind#10 @dfm
I'm not sure I totally follow the argument about how crucial it is to define both a JVP and VJP. I think that the "linear" point would also hold for an implementation using `custom_jvp`?

Regardless, I think it would be great for the docs to make a very clear case about this comparison!

Side note: as far as I can tell, it's not too hard to define a custom primitive that uses the external callback mechanism under the hood (e.g. lowering with `mlir.lower_fun`), so it's possible that the compiled component of JAXBind could be replaced by `jax.pure_callback`, while keeping the same front end. This could potentially reduce the maintenance burden.
