### 2024-04-01 6772c18
jaxbind: tweak docstrings
