+++ jaxbind/jaxbind.py
@@ -43,7 +43,7 @@ FunctionType = Union[LinearFunction, MultiLinearFunction, NonLinearFunction]
 
 
 def _exec_abstract(*args, _func: FunctionType, **kwargs):
-    """JAXbind internal: bridging the user abstract function to JAX
+    """bridging the user abstract function to JAX
 
 
     Parameters
@@ -84,7 +84,7 @@ _dtype_dict = {
 
 
 def _lowering(ctx, *args, _func: FunctionType, _platform="cpu", **kwargs):
-    """JAXbind internal: lowering the user function to JAX/XLA
+    """lowering the user function to JAX/XLA
 
 
     Parameters
@@ -159,7 +159,7 @@ def _lowering(ctx, *args, _func: FunctionType, _platform="cpu", **kwargs):
 
 
 def _explicify_zeros(x):
-    """JAXbind internal: helper function instantiating array with zeros for any
+    """helper function instantiating array with zeros for any
     ad.Zero in x.
     """
     if isinstance(x, (tuple, list)):
@@ -168,15 +168,15 @@ def _explicify_zeros(x):
 
 
 def _jvp(args, tangents, *, _func: FunctionType, **kwargs):
-    """JAXbind internal: bridging the user derivative function to JAX
+    """bridging the user derivative function to JAX
 
 
     Parameters
     ----------
     args : tuple of arrays
-        Position at which the jacobian is computed.
+        Position at which the Jacobian is computed.
     tangents : tuple of arrays
-        Tangents to which the jacobian is applied.
+        Tangents to which the Jacobian is applied.
     _func : FunctionType
         namedtuple containing all arguments from 'get_linear_call' and
         'get_nonlinear_call'.
@@ -254,7 +254,7 @@ def _jvp(args, tangents, *, _func: FunctionType, **kwargs):
 # and not as a tuple as for _jvp. Thus we need *args since we don't know
 # the number of arguments.
 def _transpose(cotangents, *args, _func: FunctionType, **kwargs):
-    """JAXbind internal: bridging the user transpose function to JAX
+    """bridging the user transpose function to JAX
 
 
     Parameters
@@ -262,7 +262,7 @@ def _transpose(cotangents, *args, _func: FunctionType, **kwargs):
     cotangents : tuple of arrays
         Cotangents for the transpose/ vjp application.
     *args : tuple of arrays
-        Position at which the jacobian is computed.
+        Position at which the Jacobian is computed.
     _func : FunctionType
         namedtuple containing all arguments from 'get_linear_call' and
         'get_nonlinear_call'.
@@ -331,7 +331,7 @@ def _transpose(cotangents, *args, _func: FunctionType, **kwargs):
 
 
 def _batch(args, in_axes, *, _func: FunctionType, **kwargs):
-    """JAXbind internal: adding batching support
+    """adding batching support
 
 
     Parameters
@@ -400,8 +400,8 @@ for platform in ["cpu", "gpu"]:
 
 
 def _call(*args, _func: FunctionType, **kwargs):
-    """JAXbind internal: helper function evaluating the JAX primitve for the
-     function 'f' in _func for given *args and **kwargs.
+    """helper function evaluating the JAX primitive for the function 'f' in
+    _func for given *args and **kwargs.
     """
     return _prim.bind(*args, **kwargs, _func=_func)
 
@@ -421,7 +421,7 @@ def get_linear_call(
     Parameters
     ----------
     f, f_T : linear function respectively its transpose
-        The function signature must be (out, args, kwargs_dump), where
+        The function signature must be `(out, args, kwargs_dump)`, where
         out and args are tuples. The results of the functions should be written as
[…]
