### 2024-04-01 b01a861
docs: generate docs with sphinx
