+++ docs/source/index.rst
@@ -1,0 +1,9 @@
+JAXbind Manual
+==============
+
+Welcome to the JAXbind documentation!
+
+.. toctree::
+   :maxdepth: 1
+
+   API reference <mod/jaxbind>
