### 2024-04-01 b3c79f0
github/workflows: add pages workflow
