### 2024-04-01 42a7c79
pages: fix dependency installation
