### 2024-04-01 5ce46e0
docs/generate: write docs to _site/
