+++ docs/generate.sh
@@ -4,4 +4,4 @@ set -e
 
 sphinx-apidoc -e -o docs/source/mod jaxbind
 # sphinx-build -b html docs/source/ docs/build/
-sphinx-build -b html docs/source/ .
+sphinx-build -b html docs/source/ _site/
