### 2024-04-01 bf299b9
jaxducc0: add docsting to wgridder
