### 2024-04-01 NIFTy-PPL/JAXbind#10 @Edenhofer
For the kind of optimizations we usually do with our models ([arxiv:1901.11033](https://arxiv.org/abs/1901.11033) and [arxiv:2105.10470](https://arxiv.org/abs/2105.10470)), we need both the JVP and the VJP. Together they serve as a positive semidefinite estimator of the local "curvature" which we use in a second order optimization.

Unfortunately, `custom_jvp` is not enough. It allows you to define a custom JVP but you must rely on JAX automatically deriving the VJP for you then. For example, say you want to wrap a host-callback to the FFT, then you can define the JVP to be the FFT itself but you can not define the VJP anymore. JAX will try to derive it for you by transposing the JVP but will fail because it can not transpose a host-callback. Custom transpositions in JAX would solve exactly this problem but so far they have various limitations that make them unsuitable for the types of problems we wanted to solve (see e.g. https://github.com/google/jax/issues/13298 and https://github.com/google/jax/issues/13283).

Very interesting idea! We started with C++ as we were  wrapping individual C++ functions before generalizing the code. Looking back at it, we could probably replace all of the C++ code with something along the lines of this https://github.com/google/jax/discussions/17875 for `jax.pure_callback`. Having said that, we haven't internally settled the debate yet whether we might want to expose a version of JAXbind free of any python GIL for C/C++ which would require our custom C++ backend.
