### 2024-04-01 NIFTy-PPL/JAXbind#10 @dfm
That all makes sense, thanks! Wrt the JOSS review, I'd just say that all this should be discussed clearly in the docs. Otherwise all sounds good.
