### 2024-04-01 0dc693f
docs: add demos to doc pages
