+++ docs/source/conf.py
@@ -7,7 +7,7 @@ extensions = [
     'sphinx.ext.imgmath',     # Render math as images
     'sphinx.ext.viewcode',    # Add links to highlighted source code
     'sphinx.ext.intersphinx', # Links to other sphinx docs (mostly numpy)
-    # 'myst_parser',            # Parse markdown
+    'myst_parser',            # Parse markdown
 ]
 master_doc = 'index'
 
