+++ docs/source/index.rst
@@ -6,4 +6,5 @@ Welcome to the JAXbind documentation!
 .. toctree::
    :maxdepth: 1
 
+   Demos <demos/index>
    API reference <mod/jaxbind>
