+++ docs/source/demos/index.rst
@@ -1,0 +1,11 @@
+JAXbind Demos
+=============
+
+Welcome to the JAXbind demos!
+
+.. toctree::
+   :maxdepth: 1
+
+   01_linear_function <01_linear_function>
+   02_multilinear_function <02_multilinear_function>
+   03_nonlinear_function <03_nonlinear_function>
