+++ .github/workflows/pages.yml
@@ -28,7 +28,10 @@ jobs:
         uses: actions/configure-pages@v5
       - name: Build with sphinx
         run: |
-          pip install sphinx pydata-sphinx-theme
+          sudo apt-get install jupyter-nbconvert
+          pip install --upgrade pip ipython ipykernel
+          ipython kernel install --name "python3" --user
+          pip install sphinx jupytext pydata-sphinx-theme myst-parser
           bash docs/generate.sh
       - name: Upload artifact
         uses: actions/upload-pages-artifact@v3
