+++ docs/source/index.rst
@@ -3,6 +3,10 @@ JAXbind Manual
 
 Welcome to the JAXbind documentation!
 
+The existing interface in JAX for connecting custom code requires deep knowledge of JAX and its C++ backend.
+The aim of `JAXbind` is to drastically lower the burden of connecting custom functions implemented in other programming languages to JAX.
+Specifically, `JAXbind` provides an easy-to-use Python interface for defining custom, so-called JAX primitives supporting any JAX transformations.
+
 .. toctree::
    :maxdepth: 1
 
