### 2024-04-01 c51d3db
jaxbind/misc: add docstring to load_kwargs
