+++ jaxbind/misc.py
@@ -5,6 +5,22 @@
 
 
 def load_kwargs(kwargs_dump, /):
+    """deserialize keyword arguments
+
+    Parameters
+    ----------
+    kwargs_dump : numpy.ndarray
+        Keyword arguments serialized by JAXbind.
+
+    Returns
+    -------
+    dict : Dictionary containing the keyword arguments.
+
+    Notes
+    -----
+    - The usage of `load_kwargs` is easiest to understand by looking at the
+      demos.
+    """
     import pickle
     import numpy as np
 
