### 2024-04-01 NIFTy-PPL/JAXbind#12 @roth-jakob
Yes, I agree, just having docstrings is a bit lazy. With #13 we export the documentation as a website. Here is the URL https://nifty-ppl.github.io/JAXbind/index.html
