### 2024-04-02 8188937
docs: link to github and pypi
