### 2024-04-02 NIFTy-PPL/JAXbind#12 @dfm
This looks good to me! I think that a bit more expository technical text setting the stage for what the different parts of the API do (related to discussion in #10) would be helpful, but I think this is sufficient for the JOSS review.
