### 2024-04-02 NIFTy-PPL/JAXbind#12 @Edenhofer
Agreed, I'll amend the documentation in this direction. I will probably only be able to do so in a couple of days though (hopefully less than 2 weeks) due to time constraints.
