### 2024-04-02 #6532 @danielskatz
👋 @hawkinsp & @wsmoses - This is just a quick ping to ask you to use the command `@editorialbot generate my checklist` to create your review checklist, and to check off the CoI and CoC items, just to make sure all the permissions are working correctly.. @editorialbot commands need to be the first thing in a new comment.  Thanks!
