### 2024-04-02 NIFTy-PPL/JAXbind#12 @roth-jakob
Thanks! Let's merge #13 first and work on #10 in a separate PR.
