+++ docs/source/conf.py
@@ -1,15 +1,15 @@
 import jaxbind
 
-needs_sphinx = '3.2.0'
+needs_sphinx = "3.2.0"
 
 extensions = [
-    'sphinx.ext.napoleon',    # Support for NumPy and Google style docstrings
-    'sphinx.ext.imgmath',     # Render math as images
-    'sphinx.ext.viewcode',    # Add links to highlighted source code
-    'sphinx.ext.intersphinx', # Links to other sphinx docs (mostly numpy)
-    'myst_parser',            # Parse markdown
+    "sphinx.ext.napoleon",  # Support for NumPy and Google style docstrings
+    "sphinx.ext.imgmath",  # Render math as images
+    "sphinx.ext.viewcode",  # Add links to highlighted source code
+    "sphinx.ext.intersphinx",  # Links to other sphinx docs (mostly numpy)
+    "myst_parser",  # Parse markdown
 ]
-master_doc = 'index'
+master_doc = "index"
 
 myst_enable_extensions = [
     "amsmath",
@@ -18,13 +18,14 @@ myst_enable_extensions = [
     "tasklist",
 ]
 
-intersphinx_mapping = {"numpy": ("https://numpy.org/doc/stable/", None),
-                       #"matplotlib": ('https://matplotlib.org/stable/', None),
-                       "ducc0": ("https://mtr.pages.mpcdf.de/ducc/", None),
-                       "scipy": ('https://docs.scipy.org/doc/scipy/reference/', None),
-                       }
+intersphinx_mapping = {
+    "numpy": ("https://numpy.org/doc/stable/", None),
+    # "matplotlib": ('https://matplotlib.org/stable/', None),
+    "ducc0": ("https://mtr.pages.mpcdf.de/ducc/", None),
+    "scipy": ("https://docs.scipy.org/doc/scipy/reference/", None),
+}
 
-autodoc_default_options = {'special-members': '__init__'}
+autodoc_default_options = {"special-members": "__init__"}
 
 napoleon_google_docstring = False
 napoleon_numpy_docstring = True
@@ -36,9 +37,9 @@ napoleon_include_special_with_doc = True
 
 imgmath_embed = True
 
-project = u'JAXbind'
-copyright = u'2024, Max-Planck-Society'
-author = u'Jakob Roth, Martin Reinecke, Gordian Edenhofer'
+project = "JAXbind"
+copyright = "2024, Max-Planck-Society"
+author = "Jakob Roth, Martin Reinecke, Gordian Edenhofer"
 
 release = jaxbind.__version__
 version = release[:-2]
@@ -61,13 +62,13 @@ html_theme_options = {
             "name": "Github",
             "url": "https://github.com/NIFTy-PPL/JAXbind",
             "icon": "fab fa-github",
-        }
+        },
     ],
     "navbar_persistent": ["search-field"],
     "navbar_end": ["navbar-icon-links"],
 }
 
 
-html_last_updated_fmt = '%b %d, %Y'
+html_last_updated_fmt = "%b %d, %Y"
 
-exclude_patterns = ['mod/modules.rst']
+exclude_patterns = ["mod/modules.rst"]
