### 2024-04-02 5fce5f6
jaxducc0: Add SHT docstring
