+++ jaxbind/contrib/jaxducc0.py
@@ -166,6 +166,25 @@ _hp_sht = get_linear_call(
 
 
 def get_healpix_sht(nside, lmax, mmax, spin, nthreads=1):
+    """Create a JAX primitive for the ducc0 SHT synthesis for HEALPix
+
+    Parameters
+    ----------
+    nside : int
+        Parameter of the HEALPix sphere.
+    lmax, mmax : int
+        Maximum l respectively m moment of the transformation (inclusive).
+    spin : int
+        Spin to use for the transfomration.
+    nthreads : int
+        Number of threads to use for the computation. If 0, use as many threads
+        as there are hardware threads available on the system.
+
+    Returns
+    -------
+    op : JAX primitive
+        The Jax primitive of the SHT synthesis for HEALPix.
+    """
     hpxparam = ducc0.healpix.Healpix_Base(nside, "RING").sht_info()
 
     hpp = partial(
@@ -264,7 +283,7 @@ def get_wgridder(
 
     Returns
     -------
-    op : Jax primitive evaluating the ducc wgridder.
+    op : JAX primitive evaluating the ducc wgridder.
         The Jax primitive has the
         signature `(uvw, freq, image)` with `uvw` being an (N, 3) array the uvw
         coordinates of the visibilities in meter, `freq`  a 1D array with the
