### 2024-04-02 d879748
jaxducc0: Add nalm docstring
