### 2024-04-08 ac84cb9
Document design choices
Resolves #10 .
