### 2024-04-08 615b3e1
internals.rst: Fix punctuation error
