### 2024-04-08 04b5fb2
internals.rst: Fix RST syntax for triton link
