### 2024-04-08 0f59fae
internals.rst: Fix typo
Thanks @Jakob for the feedback!
