### 2024-04-08 ec91d65
internals.rst: Capitalize python
Thanks @Jakob for the feedback!
