### 2024-04-22 NIFTy-PPL/JAXbind#18 @dfm
[JOSS 6532] Clarify what "any function" means
This is related to some of the other issues that I've raised, and I think the docs are getting better, but I think there should be a bit more clarity at the top of the README and on the first page of the docs to better motivate what this module does. I'd say that the key point here is that it allows you to expose any functions that can be called from **Python**, even if it's not written using JAX primitives itself. Right now I feel like the docs aren't quite specific enough about what the use case is here.

ref: https://github.com/openjournals/joss-reviews/issues/6532
