### 2024-04-22 NIFTy-PPL/JAXbind#19 @dfm
[JOSS 6532] Expand and clarify contribution guidlines
One of the JOSS review requirements is that submissions have clear contribution guidelines, but I couldn't find them documented for JAXBind. This should probably include developer docs and info about issue/PR expectations, for example.

ref: https://github.com/openjournals/joss-reviews/issues/6532
