### 2024-04-22 9c52389
README.md: Add short contributing guidelines
Resolves #19 .
