### 2024-04-22 NIFTy-PPL/JAXbind#19 @Edenhofer
Thanks for your comment! Is #20 sufficient or shall the contributions guide be more detailed?
