### 2024-04-22 d116b05
explicitly state usecase of JAXbind in comparison to external callback
