### 2024-04-22 NIFTy-PPL/JAXbind#18 @roth-jakob
Thanks for the feedback! With PR #21 we try to make this clearer. Please let us know if you think further clarification on the use case of JAXbind is needed.
