### 2024-04-22 1549243
paper.md: clarify goal of JAXbind
