### 2024-04-22 NIFTy-PPL/JAXbind#19 @dfm
I'd say that that does satisfy the letter of the JOSS requirements, but perhaps doesn't present the most welcoming face to new developers. I leave it up to you to decide if this is sufficient for your purposes!
