### 2024-04-22 NIFTy-PPL/JAXbind#18 @dfm
#21 looks good to me - thanks!
