### 2024-04-23 99e7eaa
README.md: Tweak contributing section
