### 2024-04-23 NIFTy-PPL/JAXbind#19 @Edenhofer
Thanks! You're right. The section reads really grumpy. I probably woke up on the wrong side of the bed yesterday :grimacing:

Hope it reads much friendlier now :)
