### 2024-05-05 NIFTy-PPL/JAXbind#18 @dfm
I wanted to comment here to note that the JOSS paper should also include some of this updated language. The paper text is still a bit vague on these points. I'm happy to open a separate issue if that would be useful.
