### 2024-05-05 NIFTy-PPL/JAXbind#10 @dfm
Another follow up on this. The paper has a paragraph that starts "To the best of our knowledge no other code currently exists for connecting generic functions to JAX. ..." I'd argue that this is no longer true given this discussion here. In particular, the JAX core library provides an interface for this! JAXbind does provide some extra features, but I think this should be clarified in the text.
