### 2024-05-06 eecce4e
paper: update summary according to updates in README
