### 2024-05-06 NIFTy-PPL/JAXbind#10 @roth-jakob
Thanks for the comment!

It might be that you looked at an old version of the paper. Unfortunately, I forgot to update the PDF in the JOSS review issue. Most of the changes were already incorporated into the paper. Additionally, I have now updated the summary section of the paper with PR #22.

Let me know if you believe the relation to the external callback functionally is still not clear in the paper.
