### 2024-05-06 NIFTy-PPL/JAXbind#18 @roth-jakob
Thanks for your comment!

You might have read the old PDF version from the JOSS issues. I have now updated the PDF in JOSS. Sorry that I forgot about it earlier.

Additionally, with PR #22, we have updated the summary section of the paper. Please let me know if anything is still unclear.
