### 2024-05-14 NIFTy-PPL/JAXbind#24 @wsmoses
[JOSS 6532] Case Study
You do already have the example primitive written in the paper. However, I think it would be more compelling and/or easier to use if you also included an end-to-end example demonstrating an application which required the use of jaxbind. For example binding to a python odesolver [which calls c++ under the hood].

I think this would be quite valuable for users to see how they might apply it to their application (beyond the abstract evaluation individual description in the picture).

For now I'm going to wait on marking the checkbox
" Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)." pending a response here.

ref: https://github.com/openjournals/joss-reviews/issues/6532
