+++ paper/paper.bib
@@ -13,7 +13,7 @@
   year    = {2024}
 }
 @software{Moses2024,
-  author  = {EnzymeAD},
+  author  = {Moses, William S. and Zinenko, Oleksandr},
   title   = {{Enzyme-JAX}},
   url     = {https://github.com/EnzymeAD/Enzyme-JAX},
   version = {0.0.6},
