### 2024-05-14 NIFTy-PPL/JAXbind#23 @hawkinsp
I'd second this. I think the key thing JAXbind is offering is better ergonomics than other competing approaches to extending JAX, as things stand.  That's worth highlighting and something to be proud of.
