### 2024-05-14 NIFTy-PPL/JAXbind#23 @roth-jakob
Thank you for your comments! I fully agree with this point. We are aware that there are multiple options for binding custom code to JAX. With PR #25 , we follow your suggestion, emphasizing that JAXbind provides an easy-to-use interface.

And sorry for messing up with the Enzyme-JAX author field in the bib file.
