### 2024-05-14 NIFTy-PPL/JAXbind#24 @roth-jakob
Thanks for your comment! In the paper, we believe it makes sense to keep the example simple as it makes it easier to understand the concepts. The paper is also already on the long side for the JOSS standards. In the GitHub repository, we have additional examples in the [demos](https://github.com/NIFTy-PPL/JAXbind/tree/main/demos) folder and on the [webpage](https://nifty-ppl.github.io/JAXbind/demos/index.html) .

In [jaxbind/contrib](https://github.com/NIFTy-PPL/JAXbind/tree/main/jaxbind/contrib), we provide JAX bindings for some of the functionality of the [DUCC](https://gitlab.mpcdf.mpg.de/mtr/ducc) library. Specifically, we provide bindings for Hartly transformations, Fourier transformations, spherical harmonics transformations, and the wgridder (a transformation used in radio interferometry). To make these real-world applications of JAXbind more accessible for users, we add comments to the file with PR #26  and link to it as a real-world example in the README. Additionally, with PR, we export this file to the webpage, similar to what we do for the demos.
