### 2024-05-15 ec2a495
docs.generate: streamline
