+++ docs/generate.sh
@@ -9,7 +9,6 @@ for file in "${FOLDER}/01_linear_function" "${FOLDER}/02_multilinear_function" "
     jupyter-nbconvert --to markdown --execute --ExecutePreprocessor.timeout=None "${file}.ipynb"
 done
 
-FOLDER=docs/source/jaxducc
 jupytext --to ipynb "docs/source/contrib/jaxducc0.py"
 jupyter-nbconvert --to markdown "docs/source/contrib/jaxducc0.ipynb"
 
