+++ docs/source/contrib/jaxducc0.py
@@ -1,0 +1,1 @@
+../../../jaxbind/contrib/jaxducc0.py
