+++ docs/source/contrib/index.rst
@@ -1,0 +1,10 @@
+JAXbind DUCC bindings
+=====================
+
+JAXbind provides JAX bindings for some of the functions of `DUCC <https://gitlab.mpcdf.mpg.de/mtr/ducc>`_.
+Alongside with the `demos <https://nifty-ppl.github.io/JAXbind/demos/index.html>`_ these bindings provide examples on the usage of JAXbind.
+
+.. toctree::
+   :maxdepth: 1
+
+   jaxducc0 <jaxducc0>
