### 2024-06-03 NIFTy-PPL/JAXbind#28 @hawkinsp
[JOSS 6532] Improve related work
Something that I think the related work section should at least discuss is how this work relates to the mechanisms for binding foreign functions into PyTorch and TensorFlow. It's probably worth at least a paragraph describing similarities and differences.
