### 2024-06-03 #6532 @hawkinsp
Sorry, I've been short of time. The paper LGTM, with the exception of wanting a slightly better coverage of related work (https://github.com/NIFTy-PPL/JAXbind/issues/28).
