### 2024-06-03 NIFTy-PPL/JAXbind#28 @danielskatz
For JOSS review https://github.com/openjournals/joss-reviews/issues/6532
