### 2024-06-03 NIFTy-PPL/JAXbind#28 @roth-jakob
Thanks for the feedback! I agree that this might be interesting to some readers. I will look into the TensorFlow and PyTorch functionality.
