### 2024-06-04 NIFTy-PPL/JAXbind#23 @wsmoses
Hm I'd lke to reopen this. I'm not sure my concerns about the representation of Enzyme-JaX has been addressed yet.

Per above "Similarly, Enzyme-JaX permits binding arbitrary programs, not just C++ as external code into JaX. In particular it can bind arbitrary LLVM/MLIR, including C++, and code JaX itself. It also can automatically generate the derivative of these, but similarly you can import a custom derivative as well."

I think perhaps changing "one to differentiate a C++ function" to "one to import and optionally differentiated LLVM/MLIR-based programs, like C++" would help.  Regarding "it enforces the use of Enzyme for deriving derivatives and does not allow for connecting arbitrary code to JAX." perhaps striking "enforces the use of Enzyme for defining derivatives" and changing "does not allow for connecting arbitrary code to JAX"  to "does not allow connecting arbitrary Python code to JaX".
