### 2024-06-04 #6532 @wsmoses
Made one more comment on the other related work issue here: https://github.com/NIFTy-PPL/JAXbind/issues/23#issuecomment-2146397275 but otherwise LGTM
