### 2024-06-04 NIFTy-PPL/JAXbind#23 @danielskatz
For JOSS review https://github.com/openjournals/joss-reviews/issues/6532
