### 2024-06-04 NIFTy-PPL/JAXbind#23 @Edenhofer
Yes, the wording isn't optimal yet. Maybe a much simpler statement as proposed in #29 just saying what Enzyme-JAX does is sufficient. What do you think?
