### 2024-06-04 NIFTy-PPL/JAXbind#28 @roth-jakob
With #30, we add a short paragraph about the PyTorch and TensorFlow interfaces for custom functions. Is this what you had in mind?
