### 2024-06-10 #6532 @danielskatz
@wsmoses - it looks like there's a comment for you in https://github.com/NIFTy-PPL/JAXbind/issues/23#issuecomment-2146805654
