### 2024-06-15 NIFTy-PPL/JAXbind#23 @wsmoses
Yup that sounds good to me!
