### 2024-06-19 #6532 @danielskatz
@wsmoses - just wondering again: Can you now check off the remaining items on [your checklist](https://github.com/openjournals/joss-reviews/issues/6532#issuecomment-2060481983)?
