+++ paper/paper.bib
@@ -68,7 +68,7 @@
 }
 @misc{Edenhofer2023NIFTyRE,
   author        = {{Edenhofer}, Gordian and {Frank}, Philipp and {Roth}, Jakob and {Leike}, Reimar H. and {Guerdi}, Massin and {Scheel-Platz}, Lukas I. and {Guardiani}, Matteo and {Eberle}, Vincent and {Westerkamp}, Margret and {Enßlin}, Torsten A.},
-  title         = {{Re-Envisioning Numerical Information Field Theory (NIFTy.re): A Library for Gaussian Processes and Variational Inference}},
+  title         = {Re-Envisioning Numerical Information Field Theory ({NIFTy.re}): A Library for {G}aussian Processes and Variational Inference},
   keywords      = {Software},
   year          = 2024,
   eprint        = {2402.16683},
@@ -92,6 +92,6 @@ booktitle = {29th ACM International Conference on Architectural Support for Prog
 doi = {10.1145/3620665.3640366},
 month = apr,
 publisher = {ACM},
-title = {{PyTorch 2: Faster Machine Learning Through Dynamic Python Bytecode Transformation and Graph Compilation}},
+title = {{PyTorch} 2: Faster Machine Learning Through Dynamic {P}ython Bytecode Transformation and Graph Compilation},
 url = {https://pytorch.org/assets/pytorch2-2.pdf},
 year = {2024}}
