### 2024-06-19 #6532 @danielskatz
@roth-jakob - I've suggested a number of small changes in https://github.com/NIFTy-PPL/JAXbind/pull/32. Please merge this, or let me know what you disagree with. And please do carefully check the changes, as I'm not completely sure of all of them.

In addition, there are three places where you use "respectively" that aren't understand to me. One example is "In the following, we will outline which functions our package respectively JAX requires to enable automatic differentiation" where I don't understand the intended meaning well enough to suggest an alternative.  This is also the case in "Specifically for a linear function $f$, the pushforward respectively the `jvp` of $f$ is identical to $f$ itself and independent of the position at which it is computed" and "Analogously, the pullback respectively the `vjp` becomes independent of the initial position and is given by the linear transpose of $f$, thus $(\partial f(x))^{T}(dy) = f^T(dy)$"

I also don't understand the sentence "In the future, GPU support could be added analogous to the CPU backend"
