### 2024-06-20 c5fcd3f
paper.bib: Update NIFTy.re reference
