+++ paper/paper.bib
@@ -66,15 +66,17 @@
   url    = {https://gitlab.mpcdf.mpg.de/ift/resolve},
   year   = {2024}
 }
-@misc{Edenhofer2023NIFTyRE,
-  author        = {{Edenhofer}, Gordian and {Frank}, Philipp and {Roth}, Jakob and {Leike}, Reimar H. and {Guerdi}, Massin and {Scheel-Platz}, Lukas I. and {Guardiani}, Matteo and {Eberle}, Vincent and {Westerkamp}, Margret and {Enßlin}, Torsten A.},
-  title         = {{Re-Envisioning Numerical Information Field Theory (NIFTy.re): A Library for Gaussian Processes and Variational Inference}},
-  keywords      = {Software},
-  year          = 2024,
-  eprint        = {2402.16683},
-  archiveprefix = {arXiv},
-  primaryclass  = {astro-ph.IM},
-  doi           = {10.48550/arXiv.2402.16683},
+@article{Edenhofer2023NIFTyRE,
+  title     = {Re-Envisioning Numerical Information Field Theory (NIFTy.re): A Library for Gaussian Processes and Variational Inference},
+  author    = {Gordian Edenhofer and Philipp Frank and Jakob Roth and Reimar H. Leike and Massin Guerdi and Lukas I. Scheel-Platz and Matteo Guardiani and Vincent Eberle and Margret Westerkamp and Torsten A. Enßlin},
+  year      = {2024},
+  journal   = {Journal of Open Source Software},
+  publisher = {The Open Journal},
+  volume    = {9},
+  number    = {98},
+  pages     = {6593},
+  doi       = {10.21105/joss.06593},
+  url       = {https://doi.org/10.21105/joss.06593},
 }
 
 @misc{tensorflow2015,
