### 2024-06-20 4765d83
paper: tweak GPU support statement
