### 2024-06-20 #6532 @roth-jakob
@danielskatz, thank you for proofreading the paper so quickly! I have merged NIFTy-PPL/JAXbind#32. Thanks also for your other comments. We address them with NIFTy-PPL/JAXbind#34.
