### 2024-06-20 1f61b75
paper: add martin's suggestions
