+++ paper/paper.md
@@ -78,7 +78,7 @@ Custom gradients can be added to these functions.
 
 Automatic differentiation is a core feature of JAX and often one of the main reasons for using it.
 Thus, it is essential that custom functions registered with JAX support automatic differentiation.
-In the following, we will outline which functions our package and JAX requires to enable automatic differentiation.
+In the following, we will outline which functions our package requires to enable automatic differentiation via JAX.
 For simplicity, we assume that we want to connect the nonlinear function $f(x_1,x_2) = x_1x_2^2$ to JAX.
 The `JAXbind` package expects the Python function for $f$ to take three positional arguments.
 The first argument, `out`, is a `tuple` into which the results are written.
@@ -185,7 +185,7 @@ To make use of these simplifications, `JAXbind` provides a special interface for
 # Platforms
 
 Currently, `JAXbind` only supports primitives that act on CPU memory.
-In the future, GPU support could be added.
+In the future, GPU support could be added, which should work analogously to the CPU support in most respects.
 The automatic differentiation in JAX is backend agnostic and would thus not require any additional bindings to work on the GPU.
 
 # Acknowledgements
