### 2024-06-21 #6532 @roth-jakob
I have tagged the current version and made a release. The version tag is [v1.1.0](https://github.com/NIFTy-PPL/JAXbind/releases/tag/v1.1.0). The corresponding Zendo archive doi is [10.5281/zenodo.12191214](https://doi.org/10.5281/zenodo.12191214). Please let me know if we have to do anything else.
