### 2024-06-21 #6532 @danielskatz
@roth-jakob - I see the license in the archive is CC-BY - you might want to change it to match the software license, which you should be able to do in the metadata without creating a new deposit
