### 2024-06-21 #6532 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
