### 2024-06-21 #6532 @danielskatz
👋 @roth-jakob - I see some problems in the suggested citation and the presentation of the paper on the JOSS cite

Can I ask that you remove the backquotes in the title in the .md file?
Then I will proceed to reaccept the paper, which will clean up these issues.
